Only unsubscribe events when deleting a client (#5239)
### Motivation

Within a UI context event handlers are automatically unsubscribed when
the client disconnects. But this is slightly incorrect because the
client can reconnect.

### Implementation

This PR uses `on_delete` instead of `on_disconnect` to register the
unsubscription.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation has been updated.
- [x] When uncommenting the whole auto-unsubscribe block, I still can't
cause any memory leak. So do we even need this feature? --> Ah ok,
clients are still pruned, but the list of `callbacks` would grow
indefinitely. So it is good to unsubscribe.
- [x] The name of the optional parameter is still
`unsubscribe_on_disconnect` to avoid breaking the API and the
documentation still talks about client disconnection. Is this ok?