Move vbuild into our repository (#5255)
### Motivation

The latest release of vbuild doesn't support Python 3.14 (yet) even
though the problem is fixed on main.
The release of NiceGUI 3.0.4 is currently blocked because PyPI refuses
to accept a direct dependency to a Git repository.

### Implementation

Since vbuild is only a single file of Python code, this PR simply moves
it into our repository. To silence linters and type checkers, I
drastically cleaned up and refactored the code (545 lines down to 120).
I hope everything is still working, but tests are green and Single-File
Components like `ui.plotly` and `ui.joystick` are working.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.