Let pytest run first in the pipeline, startups later. (#5254)
### Motivation

There are various issues with running the `test_startup.sh` before the
`pytest`, in ascending order of seriousness:

1. Delayed to see the results of the pytest
2. Installation of the example's requirements.txt, **defeats the purpose
of having pinned dependencies in `poetry.lock`**, as we can see in
https://github.com/zauberzeug/nicegui/actions/runs/18400412564/job/52428146319
that packages indicated in `poetry.lock` are then uninstalled in place
of other ones.

We can tolerate with point 1 but point 2 definitely needs improvement. 

This could be the cause of the many errors in #5242. Hence this PR. 

**Real motivation: I squeezed pytest into 5 minutes in #5249, but I need
to wait 10 minutes for `test_startup.sh`?**

### Implementation

Move stage `test startup` to the back. 

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).