Suggest user_plugin in demo to avoid Selenium dependency (#5262)
### Motivation

We encourage devs to avoid the screen tests because they are slower and
require Selenium to be installed. But the example in the docs uses the
generic plugin which also requires Selenium.

### Implementation

Updated demo to use non-selenium user-plugin. In the description above
it is clearly explained that there are other plugins to pick. But the
demo should be as simple as possible.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation has been updated