Fix clipboard button in `ui.code` after content has been changed (#5298)
### Motivation

In #5177 we noticed that the copy button writes old content to the
clipboard when the content of `ui.code` has been updated:
```py
c = ui.code('old content')
c.content = 'new content'
```
Clicking the button will write "old content" to the clipboard.

### Implementation

This PR introduces a new "content" prop to hold the raw (non-HTML)
version of the content. When clicking the copy button, a JavaScript call
will write the value of this prop to the clipboard.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are probably not necessary (and probably pretty hard to
write).
- [x] Documentation is not necessary.