Allow preventing default handlers for key events (#5222)
### Motivation

In #4924 @samuller mentioned that the "/" key on the NiceGUI website
doesn't work flawlessly to open the search dialog on Firefox because the
default handler opens Firefox' built-in search field. We also discussed
how to bring the `preventDefault()` and `stopPropagation()` functions to
`ui.keyboard`.

### Implementation

- This PR adds the original JavaScript event object to the argument
dictionary which is passed to `js_handler`.
- A new demo shows how to use `.on('key', ..., js_handler=...)` to
handle key events on the client and conditionally call
`preventDefault()` or `stopPropagation()`.
- A similar `js_handler` function is used to improve the "/" key
behavior on the NiceGUI website.
- The formatting of some version specifiers is fixed.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation has been added.