Mount `ui.xterm` in sync to avoid `this.terminal` undefined error (#5350)
### Motivation

Fix #5343, where since the methods are sync but mount is async, we
"mount too late" and methods cannot write to the terminal.

### Implementation

Move from async mounted, into mounted. Since we're loading CSS we can
load it whenever(TM), no need to block the rest of the mount logic.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

@falkoschindler The problem is indeed solved, but is going from async
component back to sync component a breaking change?

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>