Add `show_stats` to `ui.scene` (3.0) (#5349)
### Motivation

Originally I guessed that Falko would like high-FPS `ui.scene` for the
RoSys robots for their projects, this is why he has looked specifically
at #4825 out of the many "feat" PRs that I made was the time.

So then I did #5344 which standalone the FPS parameter adjustment. 

But then, in #4825 we got away with a more concise documentation by
documenting the stat and the FPS together. Therefore I thought we should
do this in time for 3.2

### Implementation

Take the logic from #4825 and apply to here. 

- `this.show_stats` props`
- Spawn and update the element appropriately
- Change the packed JS

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary. 
- [x] Documentation has been added.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>