Fix of recursion in case of table rows modification (#5355)
### Motivation

This patch fixes invalid recursion, which is being called, when original
table rows are being used within update_rows call. That results in
infinite recursion as table rows are Observable, which triggers change
event to itself

### Implementation

Simple check to avoid any observable to push itself in list of change
handlers. Also specific test is added to showcase the issues

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

---------

Signed-off-by: Stefan Gula <steweg@gmail.com>
Co-authored-by: Falko Schindler <falko@zauberzeug.com>