Add the anti-hang slot template for the user if they have lists in the cells (#5290)
### Motivation

Fixes #2744 and #4837 for good, in which row values being lists cause
Vue to grind to a halt.

### Implementation

Contrast to #4694 this PR's approach is to:

1. Apply the slot template which ceases the hanging for the user
(https://github.com/zauberzeug/nicegui/issues/2744#issuecomment-2881607557)
2. Informs the user they better change their code or they must live with
console warnings.

Not discovering this solution earlier was caused by how the discussion
was spread in multiple places and that the comment is hidden by "show
more items"

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>