Store the `ResizeObserver` and cleanup in `unmounted` (#5374)
### Motivation

Also acknowledged in
https://github.com/zauberzeug/nicegui/pull/5369#issuecomment-3460665023,
we `new ResizeObserver` without storing a reference of that anywhere,
let alone disconnect it when unmounted.

In some linters this is even flagged as a violation:
https://eslint-react.xyz/docs/rules/web-api-no-leaked-resize-observer,
so we better fix it.

### Implementation

Store into `this.resizeObserver`, and disconnect when `unmounted()` as
recommend by Vue docs https://vuejs.org/api/options-lifecycle#unmounted

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

### Additional notes

Should not merge conflict with #5373...

---------

Co-authored-by: Falko Schindler <mail@falkoschindler.de>
Co-authored-by: Falko Schindler <falko@zauberzeug.com>