Remove `beforeDestroy` in place of `beforeUnmount` (complete the Vue 3 migration) (#5373)
### Motivation

@falkoschindler initially spotted beforeDestroy in the codebase which
should not exist since we're using Vue 3.

https://github.com/zauberzeug/nicegui/pull/5369#issuecomment-3460665023

NiceGUI v1.0.0 may not be that complete in the migration to Vue 3...
https://github.com/zauberzeug/nicegui/releases/tag/v1.0.0

### Implementation

Follow
https://v3-migration.vuejs.org/breaking-changes/vnode-lifecycle-events.html#migration-strategy

`beforeDestroy` -> `beforeUnmount` everywhere. 

For `ui.echart` and `ui.json_editor` since the functions are duplicate,
just delete. For `ui.scene` and `ui.scene_view` apply rename.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

### Additional notes

**Not sure if Vue 3 still executes `beforeDestroy`, so potentially we
may have a client-side memory leak for `ui.scene` and `ui.scene_view`**
so I think this is a **medium** or above.