Introduce pytests for `VBuild` (#5393)
### Motivation

Improvements to `VBuild` like PR #5257 are dangerous because the
implementation is currently untested.

### Implementation

This PR adds six simple test cases to verify the core functionality of
the Vue parser. A helper function `check` allows to write tests with
almost no overhead, allowing to focus on the SFC input and the
HTML/CSS/JS output.

A few minor changes in vbuild.py (whitespace only) ensures a clean test
result.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation is not necessary.