Avoid conflict between `responsive` and `resizeObserver` (#5404)
### Motivation

In #4870 we observed a jumping button in combination with a Plotly
element:
```py
ui.plotly(go.Figure())
ui.button('Change colormap', on_click=plot.update)
```
As it turns out, it is caused by Plotly's "responsive" setting which
causes the plot to shrink to size 0x0.

### Implementation

This PR removes the "responsive" setting if it is `true`. In this case
our `resizeObserver` from PR #5369 handles the change of the container
size anyway and would conflict with this setting.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.