Let pywebview create a storage path if needed (#5419)
### Motivation

In #4882 we noticed problems with NiceGUI's random storage path for
native apps based on `tempfile.mkdtemp()`. A bad path can cause
`0x8000FFFF` WebView2 errors when calling `webview.start()`.

### Implementation

This PR removes our custom storage path and relies on webview's default.
This kind of contradicts our discussion in #4805 where we hesitated to
leave it up to pywebview. But experiments show that storage works as
expected with privacy mode (the default) and without, even with multiple
app instances.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.