Modernize our `vbuild` to use ESM import rather than inlining the JS (#5257)
### Motivation

The mastermind is and always has been @xaptronic (see #4673), but things
went stale and that PR becomes harder and harder to merge as time
marches onwards.

The potential advantages are at
https://github.com/zauberzeug/nicegui/pull/4673#issue-3025392294 but you
should read above the first `---` because the rest are about the
refactoring of `vbuild` and no AI can beat @falkoschindler (look at
#5255)

### Implementation

The logic is 3:

1. Cut the middleman `create_vue_component` between `self.script` and
`parser.script` (check
https://github.com/zauberzeug/nicegui/pull/4673#issuecomment-2852867163)
2. Use `js_imports` instead of `vue_scripts`
3. Serve `vue_components` from `_get_component`

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
  - [x] The pytest should pass, obviously
- [x] We _really_ need to have our own set of pytest to test whether our
Vue integration works. We already winged it in #5255, and we don't want
to double-wing it.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).