Add HTML IDs to complex Quasar elements (#5420)
### Motivation

In #5417 we noticed that some UI elements still don't get an ID in the
HTML DOM.

### Implementation

This PR uses a new technique by setting the "for" prop in Python. This
saves us dedicated JS components for all these elements. The `ui.input`
element, which already fixed the missing ID with an ":for" attribute in
JS, has been updated for consistency.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] Documentation is not necessary.

---------

Co-authored-by: evnchn <evanchan040511@gmail.com>