Avoid clients being deleted too early (#5422)
### Motivation

This PR finally solves the KeyError described in #5119, which has been
silenced in #5090 without understanding the root cause. As it turned
out, `Client.prune_instances` falsely treated reconnecting clients as
stale and deleted them. After the client gave up waiting for a
reconnect, it got deleted as second time.

### Implementation

This PR
1. checks if there are `client._delete_tasks`, and
2. removes the condition `if self.id in Client.instances` from #5090.

This should fix the root cause and stops hiding KeyErrors in case
clients are still deleted multiple times.

We could also make `delete()` idempotent so that it can be called
repeatedly. But as this bug shows, it's better to crash if an assumption
doesn't hold than to hide it with too many precautions.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] `test_user_storage_is_pruned` is failing. Why?
- [x] Documentation is not necessary.

---------

Co-authored-by: evnchn <evanchan040511@gmail.com>