Lazily start the active link refresher task
This change starts the active-links refresher background periodic task
if and when needed, i.e. the first time a active link is created by
the user application, rather than at application startup. Thereafter it
remains running (unless the background task is reset in which case the
refresher is started again as needed). Thus for an application that
never uses active-links, no background refresher task is ever started.

This means a well crafted nicegui application (which only uses bindable
properties/dataclass etc) will not wear the overhead of an unnecessary
task running every 0.1 secs.