`headwind.css`: Quasar-Tailwind compatibitlity by disabling Tailwind (#5445)
### Motivation

Props to @Moldeador for #5444, which highlights how #5405 caused Quasar
to lose animations in fixing #5399.

Moral of the story: Do not touch Quasar internals. 

### Implementation

So, I went ahead and disabled the conflicting classes of Tailwind,
calling it `headwind.css` (headwind is opposite to tailwind, I thought).

And indeed as mentioned in
https://github.com/zauberzeug/nicegui/issues/5444#issuecomment-3522024598
I act only on classes which Quasar has an equivalent.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

### Final note

**Will this break Tailwind animations which animate from `rotate-45`
(Quasar's class) to `rotate-100` (not Quasar's class)???**

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>