Load Vue as ES module and example of building Vue SFC with Vite (#5351)
### Motivation

* This PR allows syntax like `import { ref } from 'vue'` *as well as*
the currently supported `const { ref } = Vue` in Vue components
* Notably the newly supported syntax is the one used in the official Vue
documentation
* The newly supported syntax matches that which is output when Vue SFCs
are built with Vite
   * An example of how to build SFCs with Vite is provided in this PR
* This PR additionally complements
https://github.com/zauberzeug/nicegui/pull/5257
* https://github.com/zauberzeug/nicegui/pull/4673#issue-3025392294 notes
that "to access various vue3 functions / utilities which will get a
reference to the global Vue object - one could also choose to use [vue
via ESM](https://vuejs.org/guide/quick-start#using-the-es-module-build),
but this would I think duplicate code for not really any reason" ---
this PR does not duplicate Vue, but does allow it to be used as an ESM

The changes were previously discussed at
https://github.com/zauberzeug/nicegui/discussions/5328

### Implementation

* The Vue build is switched from `vue.global.js` to
`vue.esm-browser.js`.
* A new dependency type `BuiltInLibrary` has been added with a
production/non-production path
* This dependency type is only used for Vue currently, and is
unconditionally added in `client.py` (is this the best place to put
this?)
* It doesn't use timestamping or anything, but I think it's not
necessary here.
* It's then added to the import map under the name `'vue'` (matching the
Vue documentation) inside `generate_resources`
* To support the old-style usage with the global `Vue` object, it's the
module is imported in a script tag and set as a global.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests are not necessary.
- [x] An example has been added.

---------

Co-authored-by: evnchn <evanchan040511@gmail.com>
Co-authored-by: Frankie Robertson <frankie@robertson.name>
Co-authored-by: Evan Chan <37951241+evnchn@users.noreply.github.com>
Co-authored-by: Falko Schindler <falko@zauberzeug.com>