Add possibility to perform tests without NiceGUI pytest plugins (#5380)
### Motivation

As discussed and proposed in
https://github.com/zauberzeug/nicegui/discussions/5377,
this adds a context manager `async run(ui_code) as user: ...` within the
NiceGUI testing framework.
It enables testing of UI code without configuration of the NiceGUI
plugin setup proposed in the documentation.

This can be necessary if NiceGUI is used in a setup where the pytest
configuration can't be adapted.
As discussed in https://github.com/zauberzeug/nicegui/discussions/5377,
this was necessary for UI tests required within the pyinstaller plugin
project.

It can also be helpful as another way to get started with testing in a
new user's project -- without the need to configure pytest at all for
NiceGUI, minimal API overhead in the call sites, and without hidden
plugin "magic".

An example of a test function would be: 
```python
async def test_button_click():
    def ui_code() -> None:
        ui.button('Click me', on_click=lambda: ui.notify('Hello World!'))

    async with run(ui_code) as user:
        await user.open('/')
        await user.should_see('Click me')
        user.find(ui.button).click()
        await user.should_see('Hello World!')
```

### Implementation

The implementation relies on an async context manager which executes
`ui.run` and points an httpx.AsyncClient as a `User` to it, in line with
the setup and de-setup performed in `nicegui.testing.user_plugin.user`.
The context manager exposes the `User` for UI tests.

```python
@asynccontextmanager
async def run(ui_code):
    try:
        # simulate user and keep NiceGUI fully headless for tests
        os.environ['NICEGUI_USER_SIMULATION'] = 'true'

        # don't spawn reloader/native window; don't open browser
        ui.run(ui_code, reload=False, native=False, show=False)

        async with core.app.router.lifespan_context(core.app):
            async with httpx.AsyncClient(
                transport=httpx.ASGITransport(core.app),
                base_url='http://test'
            ) as client:
                yield User(client)
    finally:
        os.environ.pop('NICEGUI_USER_SIMULATION', None)
        ui.navigate = Navigate()
        ui.notify = notify
        ui.download = download
```

Open points in the architecture would be:
- Where could this implementation live in the code base?
- Suitable name: Candidates would be `run`, `user_run` to not risk
overloading the very simple `run` later, `user_of`, ...
- Option to create multiple users for the same UI code. Could yield
multiple users in context manager if optional argument `n_users>1`, for
example.

I'm open for discussion and refine the implementation if you consider
this useful.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

---------

Co-authored-by: Rodja Trappe <rodja@zauberzeug.com>
Co-authored-by: Falko Schindler <falko@zauberzeug.com>