Check if property is None to avoid extraneous "The type of the None singleton." (#5489)
### Motivation

Fix #5488. It is best if the documentation works identically, when ran
locally VS on the official website.

But now, due to additional docstring for `None`, there is extraneous
"The type of the None singleton."

### Implementation

Check explicitly if `property_ is not None` before proceeding with the
rest of the docstring rendering

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

### Final notes

**Too simple-win for me to take a screenshot, look at
https://github.com/zauberzeug/nicegui/issues/5488 for the results
showcase**