Show a screenshot per example on the website (#5044)
### Motivation

There is an open Enhancement in NiceGUI Wishlist project [No.
10](https://github.com/orgs/zauberzeug/projects/6?pane=issue&itemId=103129975):
Show a screenshot per example on the website

PR #4391 was created but has requested changes, so I'm creating this new
PR to follow the proposed plan by @falkoschindler:

### Implementation

This PR has a sample of 5 examples with screenshots and README files to
get some feedback before doing the others 45.

### Progress

- [5/50] Regenerate screenshots manually to better fit the UI content.
- [5/50] Add a README.md to every example folder, including at least: a
title,
a short description, an image pointing to the corresponding
screenshot.png
[examples.py](https://github.com/zauberzeug/nicegui/blob/main/website/examples.py)
- [0/1] Update
[examples.py](https://github.com/zauberzeug/nicegui/blob/main/website/examples.py)
to automatically load data from the examples folder, fetching title and
description from the readme files.
- [0/50] Improve the example tiles on the main page to somehow include
screenshots. Popups might be a good start.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>