Load codehilite.css faster (cache headers, file name hash cache-busting) (#5497)
### Motivation

While working on #5493 I also noticed that, full cached, NiceGUI
documentation reaches out to:

- `index.html` (obviously)
- `codehilite.css`

The latter we should be able to get rid of? 

### Implementation

- Filename may change so we do `${this.filename}` with the accompanying
props.
- Compute the file name and return the response using
`_generate_codehilite_css()`, which is shared.
- [ ] (Questionable) Apply `@lru_cache(maxsize=1)` to
`_generate_codehilite_css()` because it doesn't ever change in runtime
(an asusmption, is this right?)

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added (or are not necessary).
- [x] Documentation has been added (or is not necessary).

### Results

Slow 4G no CPU throttling (because network is the bottleneck)

Before: 1.46s load
After: 1.00s load

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>