Lazily start active link refresher task (#5438)
This is a simple change that merely starts the active-links refresher
background periodic task when needed, i.e. when there are one or more
links to be checked. When there are no active links, the task is
stopped. For an application that never uses active-links, no background
refresher task is ever started.

This means a well crafted nicegui application (which only uses bindable
properties/dataclass etc) will not wear the overhead of an unnecessary
task running every 0.1 secs.

---------

Co-authored-by: Falko Schindler <falko@zauberzeug.com>
Co-authored-by: evnchn <evanchan040511@gmail.com>