Handle 404 gracefully without re-raising exception (#5580)
### Motivation

In #5579 we noticed that a 404 (e.g. when opening the Chrome dev tools
which requests com.chrome.devtools.json) causes an exception with a long
stacktrace being printed to the console:

```py
from nicegui import ui

def root():
    ui.sub_pages({'/': lambda: ui.label('Home')})

ui.run(root)
```

### Implementation

This PR avoids calling `create_error_page` which would re-raise the HTTP
exception.
Instead it logs a warning and responds with a default error page.

### Progress

- [x] I chose a meaningful title that completes the sentence: "If
applied, this PR will..."
- [x] The implementation is complete.
- [x] Pytests have been added.
- [x] Documentation is not necessary.